+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 #### Changes
 
+* CORE; All Clients: Add `isServerless` parameter to IAM authentication for ElastiCache Serverless support ([#5248](https://github.com/valkey-io/valkey-glide/issues/5248))
 * JAVA: Add EVAL_RO, EVALSHA_RO, and SCRIPT DEBUG commands ([#5125](https://github.com/valkey-io/valkey-glide/pull/5125))
 * CORE: Add client certificate and private key support for mTLS ([#5092](https://github.com/valkey-io/valkey-glide/issues/5092))
 * Python: Add client certificate and private key support for mTLS ([5123](https://github.com/valkey-io/valkey-glide/issues/5123))